Add casts to fix unmatched signs in integer comparisons
Android compilation failed with:
external/pcsc-lite/src/simclist.c:1342:17: error: comparison of integers of different signs: 'ssize_t' (aka 'int') and 'uint32_t' (aka 'unsigned int') [-Werror,-Wsign-compare]
                READ_ERRCHECK(fd, buf, header.elemlen);

+++ src/simclist.c
@@ -1337,7 +1337,7 @@ int list_restore_filedescriptor(list_t *restrict l, int fd, size_t *restrict len
             if (NULL == buf)
                 return -1;
             for (cnt = 0; cnt < header.numels; cnt++) {
-                READ_ERRCHECK(fd, buf, header.elemlen);
+                READ_ERRCHECK(fd, buf, (ssize_t) header.elemlen);
                 list_append(l, l->attrs.unserializer(buf, & elsize));
                 totmemorylen += elsize;
             }
@@ -1347,7 +1347,7 @@ int list_restore_filedescriptor(list_t *restrict l, int fd, size_t *restrict len
                 buf = malloc(header.elemlen);
                 if (NULL == buf)
                     return -1;
-                READ_ERRCHECK(fd, buf, header.elemlen);
+                READ_ERRCHECK(fd, buf, (ssize_t) header.elemlen);
                 list_append(l, buf);
             }
             totmemorylen = header.numels * header.elemlen;
@@ -1362,7 +1362,7 @@ int list_restore_filedescriptor(list_t *restrict l, int fd, size_t *restrict len
                 buf = malloc((size_t)elsize);
                 if (NULL == buf)
                     return -1;
-                READ_ERRCHECK(fd, buf, elsize);
+                READ_ERRCHECK(fd, buf, (ssize_t) elsize);
                 totreadlen += elsize;
                 list_append(l, l->attrs.unserializer(buf, & elsize));
                 totmemorylen += elsize;
@@ -1374,7 +1374,7 @@ int list_restore_filedescriptor(list_t *restrict l, int fd, size_t *restrict len
                 buf = malloc(elsize);
                 if (NULL == buf)
                     return -1;
-                READ_ERRCHECK(fd, buf, elsize);
+                READ_ERRCHECK(fd, buf, (ssize_t) elsize);
                 totreadlen += elsize;
                 list_append(l, buf);
             }
